JuKeBOX

Documentation for JuKeBOX.

JuKeBOX.DblPower

JuKeBOX.Kerr

JuKeBOX.components

JuKeBOX.metric

JuKeBOX.radialroots

JuKeBOX.ℛ

JuKeBOX.DblPower — Type

DblPower

A double power law for a profile

JuKeBOX.Kerr — Type

struct Kerr{S}

Kerr spacetime with spin spin. We assume that we are in unitless scales where the black hole mass defines the length, so M=1.

JuKeBOX.components — Method

components(g)

Returns the components of the metric tensor g in SMatrix.

JuKeBOX.metric — Method

metric(g, t, r, θ, ϕ)

Construct the metric tensor for spacetime g at the location t, r, θ, ϕ.

Returns a MetricTensor object. To access the components of the tensor please the components function.

JuKeBOX.radialroots — Method

radialroots(λ, η, a)

Finds the radial roots using the of the geodesic with energy-scaled angular momentum λ and carter constant η.

JuKeBOX.ℛ — Method

ℛ(g, λ, η)

Computes the radial potential of the metric tensor assuming the energy scaled angular momentum λand carter constantη``